requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.4
django-appconf==1.0.6
django-decouple==2.1
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.3
django-redis==5.4.0
django-select2==8.2.3
django-unfold==0.43.0
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
redis==5.2.1
sqlparse==0.5.3
tablib==3.7.0
